time-series/forecasting_brent_oil/requirements.txt
pandas>=1.5.0
yfinance>=0.2.0
requests>=2.28.0
python-dotenv>=1.0.0
matplotlib>=3.6.0
seaborn>=0.12.0
statsmodels>=0.13.0
jupyterlab>=3.5.0
notebook>=6.5.0
scikit-learn>=1.2.0
plotly>=5.11.0